HT/api/requirements.txt
Flask==2.3.3 
flask-socketio==5.3.6 
eventlet==0.33.3 
Flask-SQLAlchemy==3.0.5 
psycopg2-binary==2.9.7 
cryptography==41.0.4 
numpy==1.24.3 
python-dotenv==1.0.0 
PyJWT==2.8.0 
Flask-CORS==4.0.0
requests==2.31.0
psutil==5.9.6
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
protobuf==4.25.1